chore(egh): move egghead lesson and course functions
